+++ consensus/optsync/src/client/start.rs
@@ -18,7 +18,7 @@ pub async fn start(
 ) {
     let mut client_network = NClient::<ClientMsg, Transaction>::new(c.root_cert.clone());
     let servers = c.net_map.clone();
-    let send_id = c.num_nodes as u16;
+    let send_id = c.num_nodes;
     let (net_send, mut net_recv) = 
         client_network.setup(servers, EnCodec::new(), Codec::new()).await;
